+++ src/main.js
@@ -16,6 +16,7 @@ import { seedHash, hash, random, randomSeed, openSimplexNoise, noiseProfile } fr
 import { PVector, Matrix, Plane, cross, rotX, rotY, trans, transpose, copyArr } from "./js/3Dutils.js";
 import { timeString, roundBits } from "./js/utils.js";
 import { texturesFunc, blockData, BLOCK_COUNT } from "./js/blockData.js";
+import { createDatabase, loadFromDB, saveToDB, deleteFromDB } from "./js/indexDB.js"
 
 window.blockData = blockData
 window.canvas = document.getElementById("overlay")
@@ -100,74 +101,6 @@ async function MineKhan() {
 		canvas.style.cursor = type
 	}
 	randomSeed(Math.random() * 10000000 | 0)
-
-	async function createDatabase() {
-		return await new Promise((resolve, reject) => {
-			let request = window.indexedDB.open("MineKhan", 1)
-
-			request.onupgradeneeded = function(event) {
-				let DB = event.target.result
-				// Worlds will contain and ID containing the timestamp at which the world was created, a "saved" timestamp,
-				// and a "data" string that's identical to the copy/paste save string
-				let store = DB.createObjectStore("worlds", { keyPath: "id" })
-				store.createIndex("id", "id", { unique: true })
-				store.createIndex("data", "data", { unique: false })
-			}
-
-			request.onsuccess = function() {
-				resolve(request.result)
-			}
-
-			request.onerror = function(e) {
-				console.error(e)
-				reject(e)
-			}
-		})
-	}
-	async function loadFromDB(id) {
-		let db = await createDatabase()
-		let trans = db.transaction("worlds", "readwrite")
-		let store = trans.objectStore("worlds")
-		let req = id ? store.get(id) : store.getAll()
-		return await new Promise(resolve => {
-			req.onsuccess = function() {
-				resolve(req.result)
-				db.close()
-			}
-			req.onerror = function() {
-				resolve(null)
-				db.close()
-			}
-		})
-	}
-	async function saveToDB(id, data) {
-		let db = await createDatabase()
-		let trans = db.transaction("worlds", "readwrite")
-		let store = trans.objectStore("worlds")
-		let req = store.put({ id: id, data: data })
-		return new Promise((resolve, reject) => {
-			req.onsuccess = function() {
-				resolve(req.result)
-			}
-			req.onerror = function(e) {
-				reject(e)
-			}
-		})
-	}
-	async function deleteFromDB(id) {
-		let db = await createDatabase()
-		let trans = db.transaction("worlds", "readwrite")
-		let store = trans.objectStore("worlds")
-		let req = store.delete(id)
-		return new Promise((resolve, reject) => {
-			req.onsuccess = function() {
-				resolve(req.result)
-			}
-			req.onerror = function(e) {
-				reject(e)
-			}
-		})
-	}
 
 	function save() {
 		saveToDB(world.id, {
